projects/my_project/requirements-dev.txt
docker
gitpython
typer
